requirements.txt
# UploadM8 Backend Requirements
# Using flexible versions with pre-built wheels (no Rust compilation needed)

# Web Framework (FastAPI 0.121+ pulls Starlette ≥0.49 for security fixes)
fastapi>=0.121.3,<0.140.0
uvicorn[standard]>=0.27.0,<0.35.0

# Database
asyncpg>=0.29.0,<0.32.0
python-multipart>=0.0.22,<0.1.0


# Redis
redis>=5.0.0,<6.0.0

# Authentication & Security
bcrypt>=4.1.0,<5.0.0
PyJWT>=2.8.0,<3.0.0
cryptography>=42.0.0,<47.0.0

# AWS S3 / Cloudflare R2
boto3>=1.34.0,<2.0.0

# HTTP Client
httpx>=0.26.0,<1.0.0

# Stripe Billing
stripe>=7.0.0,<12.0.0

# Validation - flexible versions ensure pre-built wheels
pydantic>=2.5.0,<3.0.0
email-validator>=2.1.0,<3.0.0

# Excel Export
openpyxl>=3.1.0,<4.0.0

# Environment
python-dotenv>=1.0.0,<2.0.0

# Error monitoring (optional — set SENTRY_DSN in host secret store only)
sentry-sdk>=2.0.0,<3.0.0

# Async utilities
aiofiles>=23.2.0,<25.0.0

# Date/Time
python-dateutil>=2.8.0,<3.0.0

openai>=1.0.0,<2.0.0
requests>=2.31.0,<3.0.0
geopandas>=0.14.0,<1.0.0
shapely>=2.0.0,<3.0.0

# Data analysis — used by telemetry_stage for statistical enrichment
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<3.0.0
# M8 publish-hour prior training (HistGradientBoostingRegressor on PCI aggregates)
scikit-learn>=1.3.0,<1.7.0
shap>=0.44.0,<0.49.0

# Image processing — used by thumbnail_stage for template rendering
Pillow>=10.0.0,<11.0.0
